+++ CHANGES.md
@@ -1,4 +1,17 @@
-# Version 0.9.0 - 2017-02-21
+# Version 0.10.0 - 2017-??-??
+ * recording: Support Gnome Shell screencast DBUS service. Allows recording
+   under Gnome Wayland (#33)
+ * recording: Support WebM as output format (#73)
+ * recording: Added option to not record mouse cursor
+ * ui: Add `--start`, `--stop` and `--toggle` command line parameters to control
+       the recording (partially fixes #23)
+ * ui: Add `--backend` command line parameter to manually choose recording
+       backend (`gnome-shell`, `ffmpeg` or `avcodec` for now)
+ * fix: Fixed a possible race condition that could lead to empty or broken files (#1)
+ * fix: Moving Peek partially outside the visible area does no longer break the
+        recording. Instead the recording area is clipped to the visible part (#64)
+
+# Version 0.9.1 - 2017-02-21
  * i18n: Fixed Czech, Croatian, Korean, Dutch and Chinese (Simplified) not getting installed
 
 # Version 0.9.0 - 2017-02-20
@@ -30,7 +43,7 @@
  * ui: Change button text while rendering (#24)
  * ui: Add a `--version` command line argument
  * ui: Show file choose directly after recording stops. This way
-   rendering and choosing the file take place in parallel (#30)
+       rendering and choosing the file take place in parallel (#30)
  * recording: Correctly scale recording area on HiDPI screens (#20)
  * i18n: Arabic translation
  * i18n: Catalan translation
@@ -64,9 +77,8 @@
 # Version 0.6.0 - 2016-01-28
  * ui: Removed unused auto save option from preferences dialog
  * fix: Try to always open the file manager, not the image viewer
- * general: Changed app id to com.uploadedlobster.peek due to the
-   previous using the wrong domain name by default. This also resets
-   existing settings.
+ * general: Changed app id to com.uploadedlobster.peek due to the previous using
+            the wrong domain name by default. This also resets existing settings.
  * i18n: Updated German translation
 
 # Version 0.5.0 - 2016.01.09